group player model
