+++ music_assistant/models/player.py
@@ -55,11 +55,13 @@ from music_assistant.constants import (
     CONF_ENTRY_FLOW_MODE,
     CONF_ENTRY_HIDE_PLAYER_IN_UI,
     CONF_ENTRY_HIDE_PLAYER_IN_UI_ALWAYS_DEFAULT,
+    CONF_ENTRY_HIDE_PLAYER_IN_UI_GROUP_PLAYER,
     CONF_ENTRY_HTTP_PROFILE,
     CONF_ENTRY_OUTPUT_CHANNELS,
     CONF_ENTRY_OUTPUT_CODEC,
     CONF_ENTRY_OUTPUT_LIMITER,
     CONF_ENTRY_PLAYER_ICON,
+    CONF_ENTRY_PLAYER_ICON_GROUP,
     CONF_ENTRY_SAMPLE_RATES,
     CONF_ENTRY_SMART_FADES_MODE,
     CONF_ENTRY_TTS_PRE_ANNOUNCE,
@@ -1313,3 +1315,73 @@ __all__ = [
     "PlayerSource",
     "PlayerState",
 ]
+
+
+class GroupPlayer(Player):
+    """Helper class for a (generic) group player."""
+
+    _attr_type: PlayerType = PlayerType.GROUP
+
+    @cached_property
+    def synced_to(self) -> str | None:
+        """Return the id of the player this player is synced to (sync leader)."""
+        # default implementation: groups can't be synced
+        return None
+
+    async def get_config_entries(self) -> list[ConfigEntry]:
+        """Return all (provider/player specific) Config Entries for the player."""
+        # Return all base config entries for a group player.
+        # Feel free to override but ensure to include the base entries by calling super() first.
+        # To override the default config entries, simply define an entry with the same key
+        # and it will be used instead of the default one.
+        return [
+            *BASE_CONFIG_ENTRIES,
+            CONF_ENTRY_PLAYER_ICON_GROUP,
+            # add player control entries as hidden entries
+            ConfigEntry(
+                key=CONF_POWER_CONTROL,
+                type=ConfigEntryType.STRING,
+                label=CONF_POWER_CONTROL,
+                default_value=PLAYER_CONTROL_NATIVE,
+                hidden=True,
+            ),
+            ConfigEntry(
+                key=CONF_VOLUME_CONTROL,
+                type=ConfigEntryType.STRING,
+                label=CONF_VOLUME_CONTROL,
+                default_value=PLAYER_CONTROL_NATIVE,
+                hidden=True,
+            ),
+            ConfigEntry(
+                key=CONF_MUTE_CONTROL,
+                type=ConfigEntryType.STRING,
+                label=CONF_MUTE_CONTROL,
+                # disable mute control for group players for now
+                # TODO: work out if all child players support mute control
+                default_value=PLAYER_CONTROL_NONE,
+                hidden=True,
+            ),
+            CONF_ENTRY_AUTO_PLAY,
+            # add default entries to hide player in UI and expose to HA
+            (
+                CONF_ENTRY_HIDE_PLAYER_IN_UI_ALWAYS_DEFAULT
+                if self.hidden_by_default
+                else CONF_ENTRY_HIDE_PLAYER_IN_UI_GROUP_PLAYER
+            ),
+            (
+                CONF_ENTRY_EXPOSE_PLAYER_TO_HA
+                if self.expose_to_ha_by_default
+                else CONF_ENTRY_EXPOSE_PLAYER_TO_HA_DEFAULT_DISABLED
+            ),
+        ]
+
+    async def volume_set(self, volume_level: int) -> None:
+        """
+        Handle VOLUME_SET command on the player.
+
+        :param volume_level: volume level (0..100) to set on the player.
+        """
+        # Default implementation:
+        # This will set the (relative) volume level on all child players.
+        # free to override if you want to handle this differently.
+        await self.mass.players.set_group_volume(self, volume_level)
